Create chromatic-console
* Init project configuration

Get initial configuration with lib tree and dev dependencies configuration (eslint, mocha, nodemon) and more as 'travis'.

* Add common part (module)

The common part groups all the common functions of the lib, as 'objectFunctions' or 'arrayFunctions'.

  ** isObject(value)
    A function to check if the value is an strict object (not an array).

  ** mergeObject(reference, assignable)
    A function to merge two object into the 'reference' parameter.
    The function is built to handle the cases of lib and it does not recursively parse arrays.

  ** orderObject(obj)
    A function to create new copy of object with ordered keys.

  ** copyObject(obj)
    A function to get a copy of an object.
    The function is built to handle the cases of lib and it does not copy arrays.

  ** isArrayOfStrings(array)
    A function to check if an array includes an other type than 'string'.

  ** groupStringsOfArray(array)
    A function concat consecutives strings of arrays.

* Add styles part (class)

The style part is the part that allows to modify the styles of the strings in the console (formats, colors, backgrounds).
The call need to returns an object with all the selected styles changes against the passed options.
It's called as a module function (() => new Styles()).
Expected object can change foreground colors, backgrounds colors and format styles as bold or dim.

  Options details :

  'flat':       It's make flat styles object (styles.backgrounds.red => styles.bgred).
  'complete':   It's change the numbers of styles and includes 256 colors instead of 7.
  'include':    It's allow to include a colors from list by name in case of complete = false.

* Add console part (class)

The console part is a part to update and/or get a new console module to send pretty stylized messages.

  Options details :

  'seed':       It's replace the initial console[fn] with new stylized colors.
  'external':   It's define if call need to return new console[fn] object.
  'details':    It's an object representing the details of console[fn] styles colors.
    'log':      New console.log style colors.
    'info':     New console.info style colors.
    'warn':     New console.warn style colors.
    'error':    New console.error style colors.

* Add differents test for externals items.

Test all common.js functions for app case.
Test class Styles and Console as modules function.
Need to improve test from styles.js && console.js, internals functions aren't currently tested.